+++ src/AMAS.h
@@ -95,9 +95,9 @@ public:
     uint32 GetProfessionCount(Player* player);
     uint32 GetDateUnixJoinAccount(uint32 AccountID);
     uint32 GetDateUnixJoinCharacter(uint32 PlayerGuid);
-	int8 GetCommentCount(uint64 PlayerGuid);
+	uint32 GetCommentCount(uint64 PlayerGuid);
     int8 GetFirstByteIP(std::string IP);
-	int8 GetIPCount(std::string IP, bool IsFirstByte = false);
+	uint32 GetIPCount(std::string IP, bool IsFirstByte = false);
 
     float GetAverageItemLevel(Player* player);
